pyproject.toml
[project]
name = "schedule-project"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dependencies>=7.7.1",
    "environs>=14.3.0",
    "fastapi>=0.118.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "redis[async]>=6.4.0",
    "sqlalchemy>=2.0.43",
    "sqlalchemy-utils>=0.42.0",
    "structlog>=25.4.0",
    "uvicorn>=0.37.0",
]
